refactor: deps-installer
