+++ network/fetch/src/index.ts
@@ -1,3 +1,3 @@
 export type { FetchFromRegistry } from '@pnpm/fetching-types'
 export { fetch, type RetryTimeoutOptions } from './fetch'
-export { createFetchFromRegistry, fetchWithAgent, type AgentOptions } from './fetchFromRegistry'
+export { createFetchFromRegistry, fetchWithAgent, type AgentOptions, type CreateFetchFromRegistryOptions } from './fetchFromRegistry'
