+++ pkg-manager/plugin-commands-installation/package.json
@@ -37,12 +37,12 @@
     "@pnpm/common-cli-options-help": "workspace:*",
     "@pnpm/config": "workspace:*",
     "@pnpm/config.config-writer": "workspace:*",
+    "@pnpm/config.deps-installer": "workspace:*",
     "@pnpm/constants": "workspace:*",
     "@pnpm/core": "workspace:*",
     "@pnpm/dedupe.check": "workspace:*",
     "@pnpm/deps.status": "workspace:*",
     "@pnpm/error": "workspace:*",
-    "@pnpm/fetch": "workspace:*",
     "@pnpm/filter-workspace-packages": "workspace:*",
     "@pnpm/find-workspace-dir": "workspace:*",
     "@pnpm/get-context": "workspace:*",
@@ -50,8 +50,6 @@
     "@pnpm/lockfile.types": "workspace:*",
     "@pnpm/manifest-utils": "workspace:*",
     "@pnpm/matcher": "workspace:*",
-    "@pnpm/network.auth-header": "workspace:*",
-    "@pnpm/npm-resolver": "workspace:*",
     "@pnpm/outdated": "workspace:*",
     "@pnpm/package-store": "workspace:*",
     "@pnpm/parse-wanted-dependency": "workspace:*",
