+++ pkg-manager/plugin-commands-installation/tsconfig.json
@@ -37,6 +37,9 @@
       "path": "../../config/config-writer"
     },
     {
+      "path": "../../config/deps-installer"
+    },
+    {
       "path": "../../config/matcher"
     },
     {
@@ -67,12 +70,6 @@
       "path": "../../lockfile/types"
     },
     {
-      "path": "../../network/auth-header"
-    },
-    {
-      "path": "../../network/fetch"
-    },
-    {
       "path": "../../packages/constants"
     },
     {
@@ -98,9 +95,6 @@
     },
     {
       "path": "../../pkg-manifest/write-project-manifest"
-    },
-    {
-      "path": "../../resolving/npm-resolver"
     },
     {
       "path": "../../resolving/resolver-base"
